rename to azurefunctions-extensions-bindings-blob
