+++ azurefunctions-extensions-bindings-blob/MANIFEST.in
