+++ azurefunctions-extensions-bindings-blob/azurefunctions/extensions/bindings/blob/blobClientConverter.py
@@ -3,7 +3,7 @@
 
 from typing import Any
 
-from azure.functions.extension.base import Datum, InConverter, OutConverter
+from azurefunctions.extensions.base import Datum, InConverter, OutConverter
 
 from .blobClient import BlobClient
 from .containerClient import ContainerClient
