+++ azurefunctions-extensions-bindings-blob/samples/blob_samples_blobclient/function_app.py
@@ -9,7 +9,7 @@
 import logging
 
 import azure.functions as func
-import azure.functions.extension.blob as blob
+import azurefunctions.extensions.bindings.blob as blob
 
 app = func.FunctionApp(http_auth_level=func.AuthLevel.ANONYMOUS)
