+++ azurefunctions-extensions-bindings-blob/samples/blob_samples_blobclient/requirements.txt
@@ -3,4 +3,4 @@
 # Manually managing azure-functions-worker may cause unexpected issues
 
 azure-functions
-azure-functions-extension-blob
+azurefunctions-extensions-bindings-blob
