+++ azurefunctions-extensions-bindings-blob/samples/blob_samples_containerclient/local.settings.json
