+++ azurefunctions-extensions-bindings-blob/samples/blob_samples_storagestreamdownloader/host.json
