+++ azurefunctions-extensions-bindings-blob/samples/blob_samples_storagestreamdownloader/local.settings.json
