+++ azurefunctions-extensions-bindings-blob/tests/test_ssd.py
@@ -6,9 +6,9 @@ import unittest
 from enum import Enum
 from typing import Optional
 
-from azure.functions.extension.base import Datum
-from azure.functions.extension.blob import BlobClient, BlobClientConverter
-from azure.storage.blob import BlobClient as BlobClientSdk
+from azurefunctions.extensions.base import Datum
+from azurefunctions.extensions.bindings.blob import BlobClientConverter, StorageStreamDownloader
+from azure.storage.blob import StorageStreamDownloader as SSDSdk
 
 
 # Mock classes for testing
@@ -64,23 +64,23 @@ class MockFunction:
         self._bindings = bindings
 
 
-class TestBlobClient(unittest.TestCase):
+class TestStorageStreamDownloader(unittest.TestCase):
     def test_input_type(self):
         check_input_type = BlobClientConverter.check_input_type_annotation
-        self.assertTrue(check_input_type(BlobClient))
+        self.assertTrue(check_input_type(StorageStreamDownloader))
         self.assertFalse(check_input_type(str))
         self.assertFalse(check_input_type(bytes))
         self.assertFalse(check_input_type(bytearray))
 
     def test_input_none(self):
         result = BlobClientConverter.decode(
-            data=None, trigger_metadata=None, pytype=BlobClient
+            data=None, trigger_metadata=None, pytype=StorageStreamDownloader
         )
         self.assertIsNone(result)
 
         datum: Datum = Datum(value=b"string_content", type=None)
         result = BlobClientConverter.decode(
-            data=datum, trigger_metadata=None, pytype=BlobClient
+            data=datum, trigger_metadata=None, pytype=StorageStreamDownloader
         )
         self.assertIsNone(result)
 
@@ -88,13 +88,13 @@ class TestBlobClient(unittest.TestCase):
         datum: Datum = Datum(value=b"string_content", type="bytearray")
         with self.assertRaises(ValueError):
             BlobClientConverter.decode(
-                data=datum, trigger_metadata=None, pytype=BlobClient
+                data=datum, trigger_metadata=None, pytype=StorageStreamDownloader
             )
 
     def test_input_empty(self):
         datum: Datum = Datum(value={}, type="model_binding_data")
-        result: BlobClient = BlobClientConverter.decode(
-            data=datum, trigger_metadata=None, pytype=BlobClient
+        result: StorageStreamDownloader = BlobClientConverter.decode(
+            data=datum, trigger_metadata=None, pytype=StorageStreamDownloader
         )
         self.assertIsNone(result)
 
@@ -113,17 +113,17 @@ class TestBlobClient(unittest.TestCase):
         )
 
         datum: Datum = Datum(value=sample_mbd, type="model_binding_data")
-        result: BlobClient = BlobClientConverter.decode(
-            data=datum, trigger_metadata=None, pytype=BlobClient
+        result: StorageStreamDownloader = BlobClientConverter.decode(
+            data=datum, trigger_metadata=None, pytype=StorageStreamDownloader
         )
 
         self.assertIsNotNone(result)
-        self.assertIsInstance(result, BlobClientSdk)
+        self.assertIsInstance(result, SSDSdk)
 
-        sdk_result = BlobClient(data=datum.value).get_sdk_type()
+        sdk_result = StorageStreamDownloader(data=datum.value).get_sdk_type()
 
         self.assertIsNotNone(sdk_result)
-        self.assertIsInstance(sdk_result, BlobClientSdk)
+        self.assertIsInstance(sdk_result, SSDSdk)
 
     def test_input_invalid_pytype(self):
         content = {
@@ -140,13 +140,13 @@ class TestBlobClient(unittest.TestCase):
         )
 
         datum: Datum = Datum(value=sample_mbd, type="model_binding_data")
-        result: BlobClient = BlobClientConverter.decode(
+        result: StorageStreamDownloader = BlobClientConverter.decode(
             data=datum, trigger_metadata=None, pytype="str"
         )
 
         self.assertIsNone(result)
 
-    def test_blob_client_invalid_creation(self):
+    def test_ssd_invalid_creation(self):
         # Create test binding
         mock_blob = MockBinding(
             name="blob", direction=MockBindingDirection.IN, data_type=None, type="blob"
@@ -174,7 +174,7 @@ class TestBlobClient(unittest.TestCase):
             ],
         )
 
-    def test_blob_client_valid_creation(self):
+    def test_ssd_valid_creation(self):
         # Create test binding
         mock_blob = MockBinding(
             name="client",
@@ -185,7 +185,9 @@ class TestBlobClient(unittest.TestCase):
 
         # Create test input_types dict
         mock_input_types = {
-            "client": MockParamTypeInfo(binding_name="blobTrigger", pytype=BlobClient)
+            "client": MockParamTypeInfo(
+                binding_name="blobTrigger", pytype=StorageStreamDownloader
+            )
         }
 
         # Create test indexed_function
